resonanceFit: rename 'massDepFit.cc/h' to 'resonanceFit.cc/h'
Finally the old name for the resonance fit is completely removed.
